FR L10n for exercise 2: Routes
This also starts fixing incorrect, hackish access to the workshopper's
`exerciseFail` method, which was a breach of encapsulation and assumed
stuff about Workshopper's internal behavior that isn't true anymore
(esp. WRT i18n).  Adjusted to use reasonably-public API only.

I also added an extra hint in the problem texts, as the hinted boilerplate
was guaranteed to fail as such.