updated readme
Updated the readme to add a little more about the project and to
include a section on getting help/discussing the exercises. Also
added myself as the lead maintainer to be responsible for said help
and maintenance going forward.

TODO: add a section in the readme about the hapijs code of conduct once
the pull request is merged https://github.com/hapijs/contrib/pull/35